feat: allow icon addition for custom asset
